+++ js/login.js
@@ -1,4 +1,4 @@
-export {logout};
+
 const db = firebase.firestore();
 const form = document.getElementById('form_login');
 let op = false;
@@ -7,27 +7,6 @@ var email;
 var password;
 var estadoCuentaf;
 /*const taskForm = document.getElementById('form_login');*/
-
-async function logout () {
-    document.addEventListener('DOMContentLoaded', async (e) => {
-        e.preventDefault();
-        if (sessionStorage.getItem('estadoCuenta') == 'false') {
-            document.getElementById('login_header').innerHTML = "iniciar sesión/registrarse";
-            document.getElementById('login_header').addEventListener('click', function () {
-                window.location = './login.html';
-            });
-
-        } else {
-            document.getElementById('login_header').innerHTML = "Cerrar Sesión";
-            document.getElementById('login_header').addEventListener('click', function () {
-                sessionStorage.removeItem('idCliente');
-                sessionStorage.setItem('estadoCuenta', false);
-            });
-        }
-    })
-}
-
-addEventListener('DOMContentLoaded', logout());
 
 form.addEventListener('input', async (e) => {
     db.collection("clientes").where("emailCliente", "==", document.getElementById('emailCliente').value)
@@ -57,7 +36,7 @@ function login() {
     if (estadoCuentaf == 'true') {
         document.getElementById('login_header').innerHTML = "Cerrar Sesión";
         sessionStorage.setItem('idCliente', idCliente);
-        window.location = './registro.html';
+        window.location = './catalogo.html';
     } else {
         console.log("datos de ingreso incorrectos");
     }
