+++ registro.html
@@ -155,7 +155,7 @@
             var defaultStorage = firebase.storage();
     
         </script>
-    <script type="module" src="./js/login.js"></script>
+    <!--<script type="module" src="./js/login.js"></script>-->
     <script src="./js/registro.js"></script>
     <script src="assets/bootstrap/js/bootstrap.min.js"></script>
     <script src="https://cdnjs.cloudflare.com/ajax/libs/baguettebox.js/1.10.0/baguetteBox.min.js"></script>
